requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
arrow==1.3.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
dateparser==1.2.0
diff-match-patch==20230430
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
Faker==25.9.2
fastapi==0.116.1
fastuuid==0.12.0
filelock==3.19.1
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.9.0
googleapis-common-protos==1.70.0
griffe==0.36.9
grpcio==1.74.0
guardrails-ai==0.6.6
guardrails-api-client==0.4.0a1
guardrails_hub_types==0.0.4
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
isoduration==20.11.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain-core==0.3.17
langgraph==0.2.45
langgraph-checkpoint==2.1.1
langgraph-sdk==0.1.74
langsmith==0.1.147
lark==1.2.2
litellm==1.76.3
lxml==4.9.4
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.6.4
openai==1.102.0
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.3
ormsgpack==1.10.0
packaging==24.2
pinecone==7.3.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
propcache==0.3.2
protobuf==6.32.0
pydantic==2.11.7
pydantic_core==2.33.2
pydash==7.0.7
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-http-client==3.3.7
python-multipart==0.0.9
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==13.9.4
rpds-py==0.27.1
rstr==3.2.2
semver==3.0.4
sendgrid==6.11.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starkbank-ecdsa==2.2.0
starlette==0.47.3
tenacity==9.1.2
tiktoken==0.11.0
tokenizers==0.22.0
tqdm==4.67.1
typer==0.15.4
types-python-dateutil==2.9.0.20250822
typing-inspection==0.4.1
typing_extensions==4.15.0
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.0.7
uvicorn==0.35.0
webcolors==24.11.1
yarl==1.20.1
zipp==3.23.0
gspread==6.1.4
google-auth==2.35.0
